LayoutTests/imported/w3c/web-platform-tests/storage-access-api/requestStorageAccess-nested-cross-origin-iframe.sub.https.window-expected.txt

CONSOLE MESSAGE: Blocked access to external URL https://www.localhost:9443/storage-access-api/resources/requestStorageAccess-iframe.https.html?testCase=nested-cross-origin-frame

Harness Error (TIMEOUT), message = null
